Move meta queries into their own file. Start stubbing out AsyncMultiQuery
