+++ djangae/db/backends/appengine/commands.py
@@ -34,6 +34,7 @@ from djangae.db import constraints, utils
 from djangae.db.backends.appengine import caching
 from djangae.db.unique_utils import query_is_unique
 
+from . import meta_queries
 
 logger = logging.getLogger(__name__)
 
@@ -125,209 +126,6 @@ def log_once(logging_call, text, args):
 
 log_once.logged = set()
 
-
-def _convert_entity_based_on_query_options(entity, opts):
-    if opts.keys_only:
-        return entity.key()
-
-    if opts.projection:
-        for k in entity.keys()[:]:
-            if k not in list(opts.projection) + [POLYMODEL_CLASS_ATTRIBUTE]:
-                del entity[k]
-
-    return entity
-
-
-class QueryByKeys(object):
-    """ Does the most efficient fetching possible for when we have the keys of the entities we want. """
-
-    def __init__(self, model, queries, ordering, namespace):
-        # `queries` should be filtered by __key__ with keys that have the namespace applied to them.
-        # `namespace` is passed for explicit niceness (mostly so that we don't have to assume that
-        # all the keys belong to the same namespace, even though they will).
-        def _get_key(query):
-            result = query["__key__ ="]
-            return result
-
-        self.model = model
-        self.namespace = namespace
-
-        # groupby requires that the iterable is sorted by the given key before grouping
-        self.queries = sorted(queries, key=_get_key)
-        self.queries_by_key = { a: list(b) for a, b in groupby(self.queries, _get_key) }
-
-        self.ordering = ordering
-        self._Query__kind = queries[0]._Query__kind
-
-    def Run(self, limit=None, offset=None):
-        """
-            Here are the options:
-
-            1. Single key, hit memcache
-            2. Multikey projection, async MultiQueries with ancestors chained
-            3. Full select, datastore get
-        """
-
-        opts = self.queries[0]._Query__query_options
-        key_count = len(self.queries_by_key)
-
-        is_projection = False
-
-        results = None
-        if key_count == 1:
-            # FIXME: Potentially could use get_multi in memcache and the make a query
-            # for whatever remains
-            key = self.queries_by_key.keys()[0]
-            result = caching.get_from_cache_by_key(key)
-            if result is not None:
-                results = [result]
-                cache = False # Don't update cache, we just got it from there
-
-        if results is None:
-            if opts.projection:
-                is_projection = True # Don't cache projection results!
-
-                # Assumes projection ancestor queries are faster than a datastore Get
-                # due to lower traffic over the RPC. This should be faster for queries with
-                # < 30 keys (which is the most common case), and faster if the entities are
-                # larger and there are many results, but there is probably a slower middle ground
-                # because the larger number of RPC calls. Still, if performance is an issue the
-                # user can just do a normal get() rather than values/values_list/only/defer
-
-                to_fetch = (offset or 0) + limit if limit else None
-                additional_cols = set([ x[0] for x in self.ordering if x[0] not in opts.projection])
-
-                multi_query = []
-                final_queries = []
-                orderings = self.queries[0]._Query__orderings
-                for key, queries in self.queries_by_key.iteritems():
-                    for query in queries:
-                        if additional_cols:
-                            # We need to include additional orderings in the projection so that we can
-                            # sort them in memory. Annoyingly that means reinstantiating the queries
-                            query = Query(
-                                kind=query._Query__kind,
-                                filters=query,
-                                projection=list(opts.projection).extend(list(additional_cols)),
-                                namespace=self.namespace,
-                            )
-
-                        query.Ancestor(key) # Make this an ancestor query
-                        multi_query.append(query)
-                        if len(multi_query) == 30:
-                            final_queries.append(datastore.MultiQuery(multi_query, orderings).Run(limit=to_fetch))
-                            multi_query = []
-                else:
-                    if len(multi_query) == 1:
-                        final_queries.append(multi_query[0].Run(limit=to_fetch))
-                    elif multi_query:
-                        final_queries.append(datastore.MultiQuery(multi_query, orderings).Run(limit=to_fetch))
-
-                results = chain(*final_queries)
-            else:
-                results = datastore.Get(self.queries_by_key.keys())
-
-        def iter_results(results):
-            returned = 0
-            # This is safe, because Django is fetching all results any way :(
-            sorted_results = sorted(results, cmp=partial(utils.django_ordering_comparison, self.ordering))
-            sorted_results = [result for result in sorted_results if result is not None]
-            if not is_projection and sorted_results:
-                caching.add_entities_to_cache(
-                    self.model,
-                    sorted_results,
-                    caching.CachingSituation.DATASTORE_GET,
-                    self.namespace,
-                )
-
-            for result in sorted_results:
-                if is_projection:
-                    entity_matches_query = True
-                else:
-                    entity_matches_query = any(
-                        utils.entity_matches_query(result, qry) for qry in self.queries_by_key[result.key()]
-                    )
-
-                if not entity_matches_query:
-                    continue
-
-                if offset and returned < offset:
-                    # Skip entities based on offset
-                    returned += 1
-                    continue
-                else:
-
-                    yield _convert_entity_based_on_query_options(result, opts)
-
-                    returned += 1
-
-                    # If there is a limit, we might be done!
-                    if limit is not None and returned == (offset or 0) + limit:
-                        break
-
-        return iter_results(results)
-
-    def Count(self, limit, offset):
-        return len([x for x in self.Run(limit, offset)])
-
-
-class NoOpQuery(object):
-    def Run(self, limit, offset):
-        return []
-
-    def Count(self, limit, offset):
-        return 0
-
-
-class UniqueQuery(object):
-    """
-        This mimics a normal query but hits the cache if possible. It must
-        be passed the set of unique fields that form a unique constraint
-    """
-    def __init__(self, unique_identifier, gae_query, model, namespace):
-        self._identifier = unique_identifier
-        self._gae_query = gae_query
-        self._model = model
-        self._namespace = namespace
-
-        self._Query__kind = gae_query._Query__kind
-
-    def get(self, x):
-        return self._gae_query.get(x)
-
-    def keys(self):
-        return self._gae_query.keys()
-
-    def Run(self, limit, offset):
-        opts = self._gae_query._Query__query_options
-        if opts.keys_only or opts.projection:
-            return self._gae_query.Run(limit=limit, offset=offset)
-
-        ret = caching.get_from_cache(self._identifier, self._namespace)
-        if ret is not None and not utils.entity_matches_query(ret, self._gae_query):
-            ret = None
-
-        if ret is None:
-            # We do a fast keys_only query to get the result
-            keys_query = Query(self._gae_query._Query__kind, keys_only=True, namespace=self._namespace)
-            keys_query.update(self._gae_query)
-            keys = keys_query.Run(limit=limit, offset=offset)
-
-            # Do a consistent get so we don't cache stale data, and recheck the result matches the query
-            ret = [x for x in datastore.Get(keys) if x and utils.entity_matches_query(x, self._gae_query)]
-            if len(ret) == 1:
-                caching.add_entities_to_cache(
-                    self._model,
-                    [ret[0]],
-                    caching.CachingSituation.DATASTORE_GET,
-                    self._namespace,
-                )
-            return iter(ret)
-
-        return iter([ret])
-
-    def Count(self, limit, offset):
-        return sum(1 for x in self.Run(limit, offset))
 
 
 from djangae.db.backends.appengine.query import transform_query
@@ -630,13 +428,13 @@ class SelectCommand(object):
 
         if can_perform_datastore_get(self.query):
             # Yay for optimizations!
-            return QueryByKeys(self.query.model, queries, ordering, self.namespace)
+            return meta_queries.QueryByKeys(self.query.model, queries, ordering, self.namespace)
 
         if len(queries) == 1:
             identifier = query_is_unique(self.query.model, queries[0])
             if identifier:
                 # Yay for optimizations!
-                return UniqueQuery(identifier, queries[0], self.query.model, self.namespace)
+                return meta_queries.UniqueQuery(identifier, queries[0], self.query.model, self.namespace)
 
             return queries[0]
         else:
@@ -674,7 +472,7 @@ class SelectCommand(object):
                 # didn't seem to indicate much of a performance difference, even when doing the pk__in
                 # with GetAsync while the count was running. That might not be true of prod though so
                 # if anyone comes up with a faster idea let me know!
-                if isinstance(query, QueryByKeys):
+                if isinstance(query, meta_queries.QueryByKeys):
                     # If this is a QueryByKeys, just do the datastore Get and count the results
                     resultset = (x.key() for x in query.Run(limit=limit, offset=offset) if x)
                 else:
@@ -1115,7 +913,7 @@ class UpdateCommand(object):
                 return False
 
             if (
-                isinstance(self.select.gae_query, (Query, UniqueQuery)) # ignore QueryByKeys and NoOpQuery
+                isinstance(self.select.gae_query, (Query, meta_queries.UniqueQuery)) # ignore QueryByKeys and NoOpQuery
                 and not utils.entity_matches_query(result, self.select.gae_query)
             ):
                 # Due to eventual consistency they query may have returned an entity which no longer
